added Vue.js support and removed react dependency in index.js
