+++ __test__/Image.vue
@@ -1,0 +1,12 @@
+<template>
+  <img :src="src" />
+</template>
+<script>
+module.exports = {
+  data: () => {
+    return {
+      src: '#'
+    }
+  }
+}
+</script>
