+++ __test__/vuejs.test.js
@@ -1,0 +1,16 @@
+const { axe, toHaveNoViolations } = require('../index')
+
+expect.extend(toHaveNoViolations)
+
+const Image = require('./Image.vue')
+
+describe('Vue', () => {
+  it('renders correctly', async () => {
+
+    let results = await axe(Image)
+    
+    expect(() => {
+      expect(results).toHaveNoViolations()
+    }).toThrowErrorMatchingSnapshot()
+  })
+})
