+++ package.json
@@ -24,20 +24,37 @@
     "node": ">= 6.0.0"
   },
   "dependencies": {
+    "@vue/server-test-utils": "^1.0.0-beta.29",
+    "@vue/test-utils": "^1.0.0-beta.29",
     "axe-core": "3.1.2",
     "jest-matcher-utils": "24.0.0",
-    "lodash.merge": "^4.6.2"
+    "lodash.merge": "^4.6.2",
+    "vue-server-renderer": "^2.6.10",
+    "vue-template-compiler": "^2.6.10"
   },
   "devDependencies": {
+    "vue": "^2.6.10",
+    "vue-jest": "^3.0.4",
+    "@testing-library/react": "^8.0.5",
     "async-to-gen": "1.4.0",
+    "babel-core": "^6.26.3",
     "jest": "24.1.0",
     "react": "16.7.0",
-    "react-dom": "16.7.0",
-    "@testing-library/react": "^8.0.5"
+    "react-dom": "16.7.0"
   },
   "jest": {
+    "moduleFileExtensions": [
+      "js",
+      "json",
+      "jsx",
+      "ts",
+      "tsx",
+      "node",
+      "vue"
+    ],
     "transform": {
-      "^.+\\.js$": "async-to-gen"
+      "^.+\\.js$": "async-to-gen",
+      "^.+\\.vue$": "vue-jest"
     },
     "transformIgnorePatterns": [
       "node_modules/(?!(@testing-library/react)/)"
